Show number of deleted entries

+++ src/database/datenkatalog.rs
@@ -37,7 +37,7 @@ impl Datenkatalog {
         vec![]
     }
 
-    pub fn clean(db: &Database, id: u64) -> bool {
+    pub fn clean(db: &Database, id: u64) -> u64 {
         if let Ok(Some(name)) = "SELECT name FROM data_catalogue WHERE id LIKE :id"
             .with(params! {"id" => id})
             .first::<String, PooledConn>(db.connection())
@@ -45,14 +45,19 @@ impl Datenkatalog {
             let name_re = Regex::new(r"[[:^alpha:]]").unwrap();
             let table_name = &format!("dk_{}", name_re.replace(name.as_str(), "_").to_lowercase());
 
-            return format!(
-                "DELETE FROM prozedur WHERE id IN (SELECT id FROM {}); DELETE FROM {}",
-                table_name, table_name
-            )
-            .run(db.connection())
-            .is_ok();
+            if let Ok(Some(count)) = format!("SELECT COUNT(*) FROM {}", table_name).first(db.connection()) {
+                if format!(
+                    "DELETE FROM prozedur WHERE id IN (SELECT id FROM {}); DELETE FROM {}",
+                    table_name, table_name
+                )
+                .run(db.connection())
+                .is_ok()
+                {
+                    return count;
+                }
+            }
         };
 
-        false
+        0
     }
 }
